+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This file needs to be updated with every significant pull request. It is used to write down release notes.
 
 ## Version 0.5.6
-
+* Expose request id to parent Rack application when using `ApplicationInsights::Rack::TrackRequest` middleware through `env['ApplicationInsights.request.id']`
 
 ## Version 0.5.5
 * Add some basic logging when failed to send telemetry to the server