+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file needs to be updated with every significant pull request. It is used to write down release notes.
 
+## Version 0.5.7
+
 ## Version 0.5.6
 * Expose request id to parent Rack application when using `ApplicationInsights::Rack::TrackRequest` middleware through `env['ApplicationInsights.request.id']`.
 * Implement operation context functionality for `ApplicationInsights::Rack::TrackRequest`.